YQL-19747 Introduce name/service module
