+++ yql/essentials/sql/v1/complete/name/name_service.h
@@ -1,83 +1,3 @@
 #pragma once
 
-#include <yql/essentials/sql/v1/complete/core/statement.h>
-
-#include <library/cpp/threading/future/core/future.h>
-
-#include <util/generic/vector.h>
-#include <util/generic/string.h>
-
-namespace NSQLComplete {
-
-    using NThreading::TFuture;
-
-    struct TIndentifier {
-        TString Indentifier;
-    };
-
-    struct TNamespaced {
-        TString Namespace;
-    };
-
-    struct TKeyword {
-        TString Content;
-    };
-
-    struct TPragmaName: TIndentifier {
-        struct TConstraints: TNamespaced {};
-    };
-
-    struct TTypeName: TIndentifier {
-        using TConstraints = std::monostate;
-    };
-
-    struct TFunctionName: TIndentifier {
-        struct TConstraints: TNamespaced {};
-    };
-
-    struct THintName: TIndentifier {
-        struct TConstraints {
-            EStatementKind Statement;
-        };
-    };
-
-    using TGenericName = std::variant<
-        TKeyword,
-        TPragmaName,
-        TTypeName,
-        TFunctionName,
-        THintName>;
-
-    struct TNameRequest {
-        TVector<TString> Keywords;
-        struct {
-            std::optional<TPragmaName::TConstraints> Pragma;
-            std::optional<TTypeName::TConstraints> Type;
-            std::optional<TFunctionName::TConstraints> Function;
-            std::optional<THintName::TConstraints> Hint;
-        } Constraints;
-        TString Prefix = "";
-        size_t Limit = 128;
-
-        bool IsEmpty() const {
-            return Keywords.empty() &&
-                   !Constraints.Pragma &&
-                   !Constraints.Type &&
-                   !Constraints.Function &&
-                   !Constraints.Hint;
-        }
-    };
-
-    struct TNameResponse {
-        TVector<TGenericName> RankedNames;
-    };
-
-    class INameService {
-    public:
-        using TPtr = TAtomicSharedPtr<INameService>;
-
-        virtual TFuture<TNameResponse> Lookup(TNameRequest request) const = 0;
-        virtual ~INameService() = default;
-    };
-
-} // namespace NSQLComplete
+#include <yql/essentials/sql/v1/complete/name/service/name_service.h>
