+++ yql/essentials/sql/v1/complete/name/service/static/name_index.cpp
