+++ yql/essentials/sql/v1/complete/name/service/static/name_service.cpp
