+++ yql/essentials/sql/v1/complete/name/service/static/name_service.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "ranking.h"
+
+#include <yql/essentials/sql/v1/complete/name/service/name_service.h>
+
+namespace NSQLComplete {
+
+    struct NameSet {
+        TVector<TString> Pragmas;
+        TVector<TString> Types;
+        TVector<TString> Functions;
+        THashMap<EStatementKind, TVector<TString>> Hints;
+    };
+
+    NameSet MakeDefaultNameSet();
+
+    THolder<INameService> MakeStaticNameService();
+
+    THolder<INameService> MakeStaticNameService(NameSet names, IRanking::TPtr ranking);
+
+} // namespace NSQLComplete
