+++ yql/essentials/sql/v1/complete/name/service/ya.make
@@ -2,12 +2,10 @@ LIBRARY()
 
 PEERDIR(
     yql/essentials/sql/v1/complete/core
-    yql/essentials/sql/v1/complete/name/service
 )
 
 END()
 
 RECURSE(
-    object
-    service
+    static
 )
