+++ yql/essentials/sql/v1/complete/name/static/name_service.h
@@ -1,22 +1,3 @@
 #pragma once
 
-#include "ranking.h"
-
-#include <yql/essentials/sql/v1/complete/name/name_service.h>
-
-namespace NSQLComplete {
-
-    struct NameSet {
-        TVector<TString> Pragmas;
-        TVector<TString> Types;
-        TVector<TString> Functions;
-        THashMap<EStatementKind, TVector<TString>> Hints;
-    };
-
-    NameSet MakeDefaultNameSet();
-
-    THolder<INameService> MakeStaticNameService();
-
-    THolder<INameService> MakeStaticNameService(NameSet names, IRanking::TPtr ranking);
-
-} // namespace NSQLComplete
+#include <yql/essentials/sql/v1/complete/name/service/static/name_service.h>
