+++ yql/essentials/sql/v1/complete/sql_complete.cpp
@@ -1,7 +1,7 @@
 #include "sql_complete.h"
 
 #include <yql/essentials/sql/v1/complete/text/word.h>
-#include <yql/essentials/sql/v1/complete/name/static/name_service.h>
+#include <yql/essentials/sql/v1/complete/name/service/static/name_service.h>
 #include <yql/essentials/sql/v1/complete/syntax/local.h>
 #include <yql/essentials/sql/v1/complete/syntax/format.h>
 
@@ -51,14 +51,14 @@ namespace NSQLComplete {
         }
 
     private:
-        TCompletedToken GetCompletedToken(TStringBuf prefix) {
+        TCompletedToken GetCompletedToken(TStringBuf prefix) const {
             return {
                 .Content = LastWord(prefix),
                 .SourcePosition = LastWordIndex(prefix),
             };
         }
 
-        NThreading::TFuture<TVector<TCandidate>> GetCandidates(TLocalSyntaxContext context, const TCompletedToken& prefix) {
+        NThreading::TFuture<TVector<TCandidate>> GetCandidates(TLocalSyntaxContext context, const TCompletedToken& prefix) const {
             TNameRequest request = {
                 .Prefix = TString(prefix.Content),
                 .Limit = Configuration.Limit,
