+++ yql/essentials/sql/v1/complete/ya.make
@@ -7,8 +7,8 @@ SRCS(
 PEERDIR(
     yql/essentials/sql/v1/lexer
     yql/essentials/sql/v1/complete/antlr4
-    yql/essentials/sql/v1/complete/name
-    yql/essentials/sql/v1/complete/name/static
+    yql/essentials/sql/v1/complete/name/service
+    yql/essentials/sql/v1/complete/name/service/static
     yql/essentials/sql/v1/complete/syntax
     yql/essentials/sql/v1/complete/text
 )
